Fix up EBML lacing, it was majorly b0rked. -All lacing sizes should be the same width (was mixing 32 and 64 bit ints) -Move the calculation of the number of bytes used by sizes before BlockGroup. -Add above to the size of the BlockGroup. -Fix lacing flags. -Copy the correct number of bytes when copying the lacing sizes.
Misc: Add some comments. Rename c_delta_flags to c_delta because flags are handled
seperatly.

+++ include/libmkv.h
@@ -138,7 +138,7 @@ int  mk_writeHeader(mk_Writer *w, const char *writingApp);
 int  mk_startFrame( mk_Writer *w, mk_Track *track );
 int  mk_addFrameData(mk_Writer *w, mk_Track *track, const void *data, unsigned size);
 int  mk_setFrameFlags(mk_Writer *w, mk_Track *track, int64_t timestamp, unsigned keyframe);
-int   mk_setFrameLacing(mk_Writer *w, mk_Track *track, uint8_t lacing, uint8_t num_frames, uint32_t sizes[]);
+int   mk_setFrameLacing(mk_Writer *w, mk_Track *track, uint8_t lacing, uint8_t num_frames, uint64_t sizes[]);
 int  mk_createChapterSimple(mk_Writer *w, uint64_t start, uint64_t end, char *name);
 int  mk_close( mk_Writer *w );
